Alot of changes. Got the book, imporved lemma, el is not seperated in spanishanalyser, added io functions, e.t.c

+++ io_functions_analyse.py
@@ -4,56 +4,96 @@ import pickle
 import re
 import os
 
-def open_book(filedir):
-    with open(filedir, 'r', encoding="utf-8") as file:
+
+def open_book(file_dir):
+    with open(file_dir, 'r', encoding="utf-8") as file:
         text = file.read()
     return text
-
 
 def timing(func):
     print(timeit.timeit(func, number=1))
 
 
-def corpus_files(lang):
+def file_text(file_path):
     corpus = ""
-    for textfile in os.scandir(f"Texts/{lang}/Corpus"):
+    for textfile in os.scandir(file_path):
         with open(textfile, 'r', encoding="utf-8") as file:
             text = file.read()
         corpus += text
     return corpus
 
+def all_text(file_path, text = ''):
+    text = ''
+    for file in os.scandir(file_path):
+        if file.is_dir():
+            text += all_text(file, text)
+        else:
+            with open(file, 'r', encoding="utf-8") as txt:
+                text += txt.read()
+    return text
 
-def save(name, obj):
-    with open(f'{name}.pkl', "wb") as output:
+
+
+
+
+def save(name, obj, path='Data/Analyse_objs/'):
+    with open(f'{path}{name}.pkl', "wb") as output:
         pickle.dump(obj, output, pickle.HIGHEST_PROTOCOL)
 
 
-def load(lang):
-    with open(f'{lang}.pkl', "rb") as intput:
-        obj = pickle.load(intput)
+def load(name, path='Data/Analyse_objs/'):
+    with open(f'{path}{name}.pkl', "rb") as in_put:
+        obj = pickle.load(in_put)
     return obj
 
 
-def save_corpus(lang):
-    book = corpus_files(lang)
+def save_corpus(file_path, name, lang, path='Data/Analyse_objs'):
+    book = file_text(file_path)
     if lang == 'FR':
         obj = FrenchAnalyser(book)
     elif lang == 'Eng':
         obj = EnglishAnalyser(book)
     elif lang == 'ES':
         obj = SpanishAnalyser(book)
-    save(lang, obj)
+    save(name, obj, path)
+
+def load_corpus(name, path='Data/Analyse_objs/'):
+    base_analyser = Analyser('')
+    file_path = f'{path}/{name}'
+    for file in os.scandir(file_path):
+        with open(file, "rb") as in_put:
+            obj = pickle.load(in_put)
+            base_analyser += obj
+    return base_analyser
 
 
 # Trimming Gutenberg Additions:
-def trim(text):
+def trim(text, check=False):
+    original = text
     starts = re.compile(
         r"(?:The Project Gutenberg eBook|Project Gutenberg).{1,2000}\*\*\*\s?START OF (?:THE|THIS) PROJECT GUTENBERG EBOOK",
         flags=re.DOTALL | re.IGNORECASE)
-    endings = re.compile(r"\*\*\*\s?END OF (?:THE|THIS) PROJECT GUTENBERG EBOOK.+\*\*\*(?:[\n]|.)+")
+    endings = re.compile(r"\*\*\*\s?END OF (?:THE|THIS) PROJECT GUTENBERG EBOOK.+\*\*\*(?:[\n]|.)+",
+                         flags=re.IGNORECASE)
 
     text = starts.sub('', text)
     text = endings.sub('', text)
+
+    # checking if trimming was successful
+    if check:
+        if starts.sub('', original) == original:
+            print('First trim has failed')
+
+        else: print('First trim successful')
+
+        if endings.sub('', original) == original:
+            print('Second trim has failed')
+        else:
+            print('Second trim successful')
+
+        if text == original:
+            print('Nothing has been trimmed')
+
     return text
 
 
@@ -68,19 +108,31 @@ def trim1(text):
 
 
 def trim2(text):
-    endings = re.compile(r"\*\*\*\s?END OF (?:THE|THIS) PROJECT GUTENBERG EBOOK.+\*\*\*(?:[\n]|.)+")
+    endings = re.compile(r"\*\*\*\s?END OF (?:THE|THIS) PROJECT GUTENBERG EBOOK.+\*\*\*(?:[\n]|.)+",
+                         flags=re.IGNORECASE)
 
     text = endings.sub('', text)
     return text
 
 
-def trim_corpus(lang):
-    for textfile in os.scandir(f"Texts/{lang}/Corpus"):
+def trim_file(file_path, overwrite=False):
+    trimmed_corp = ''
+    for textfile in os.scandir(file_path):
         with open(textfile, 'r', encoding="utf-8") as file:
             txt = file.read()
             t_txt = trim(txt)
-            if txt == trim1(txt) or txt == trim2(txt):
-                print(file.name)
+            if txt == trim1(txt):
+                print(file.name, "---has failed first trim")
+            if txt == trim2(txt):
+                print(file.name, "---has failed second trim")
+            if txt == t_txt:
+                print(file.name, '---Nothing has been trimmed')
+        if overwrite:
+            with open(textfile, 'w', encoding="utf-8") as file:
+                file.write(t_txt)
 
-        with open(textfile, 'w', encoding="utf-8") as file:
-            file.write(t_txt)
+        trimmed_corp += t_txt
+
+    return trimmed_corp
+
+
